Update text to image service to generate actual image files
The text to image generation functionality of the service is now functional, as opposed to previously throwing a NotImplementedException. It pulls a directory from the service's attributes collection to use as output location for the generated images. Logging has also been improved within the service, registering status updates and error messages. Test cases are updated to match the new changes.